requirements.txt
asgiref==3.5.0
Django==3.2.12
djangorestframework==3.13.1
install==1.3.5
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
